make crc checks work on 64 bit systems, other fixes
Whether or not the tool works correctly still seems to depend on
the GCC version and/or GLIBC version used.

+++ source/dzip-zconf.h
@@ -214,7 +214,8 @@
 typedef unsigned char  Byte;  /* 8 bits */
 #endif
 typedef unsigned int   uInt;  /* 16 bits or more */
-typedef unsigned long  uLong; /* 32 bits or more */
+#include <inttypes.h>
+typedef uint32_t  uLong; /* 32 bits or more */
 
 #ifdef SMALL_MEDIUM
    /* Borland C/C++ and some old MSC versions ignore FAR inside typedef */
